requirements.txt
ultralytics
roboflow
ipython
streamlit
glob2
opencv-python-headless
pillow
numpy
torch
torchvision